Moving more phase code into shared VM.
