+++ app/src/main/java/com/adammcneilly/pocketleague/phase/ui/PhaseDetailContent.kt
@@ -4,6 +4,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.adammcneilly.pocketleague.shared.phasedetail.state.PhaseDetailViewState
 
 /**
  * Given a [viewState], render the UI for phase detail information.
