+++ shared/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/phasedetail/ui/PhaseDetailDisplayModel.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.pocketleague.phase.ui
+package com.adammcneilly.pocketleague.shared.phasedetail.ui
 
 /**
  * User friendly representation of a phase detail.
